+++ CHANGELOG.md
@@ -4,3 +4,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- ECB currency model